+++ cypress/e2e/edit-profile.cy.ts
@@ -1,3 +1,4 @@
+import encodeURIComponent from 'strict-uri-encode'
 import { Person } from '../support/commands'
 
 const profile = {