+++ cypress/e2e/map.cy.ts
@@ -1,5 +1,6 @@
 import range from 'lodash/range'
 import ngeohash from 'ngeohash'
+import encodeURIComponent from 'strict-uri-encode'
 import { Person } from '../support/commands'
 import { UserConfig } from '../support/css-authentication'
 import { AccommodationConfig } from '../support/setup'