+++ cypress/e2e/profile.cy.ts
@@ -1,3 +1,4 @@
+import encodeURIComponent from 'strict-uri-encode'
 import { Person } from '../support/commands'
 import { UserConfig } from '../support/css-authentication'