+++ src/components/PersonBadge/PersonBadge.tsx
@@ -2,7 +2,8 @@ import { Avatar } from '@/components'
 import { useConfig } from '@/config/hooks'
 import { useProfile } from '@/hooks/data/useProfile'
 import { URI } from '@/types'
-import { Link } from 'react-router-dom'
+import { Link } from 'react-router'
+import encodeURIComponent from 'strict-uri-encode'
 import styles from './PersonBadge.module.scss'
 
 export const PersonBadge = ({