+++ src/hooks/usePreviousUriAfterSolidRedirect.ts
@@ -1,6 +1,6 @@
 import { EVENTS, getDefaultSession } from '@inrupt/solid-client-authn-browser'
 import { useEffect } from 'react'
-import { useNavigate } from 'react-router-dom'
+import { useNavigate } from 'react-router'
 
 // https://docs.inrupt.com/developer-tools/javascript/client-libraries/tutorial/restore-session-browser-refresh/#use-session-restore-event-handler