+++ src/layouts/NavLayout.tsx
@@ -1,5 +1,5 @@
 import { ReactNode } from 'react'
-import { NavLink } from 'react-router-dom'
+import { NavLink } from 'react-router'
 import styles from './NavLayout.module.scss'
 
 export const NavLayout = ({