+++ src/pages/SearchHosts/AccommodationInfo.tsx
@@ -6,7 +6,8 @@ import { useAppSelector } from '@/redux/hooks'
 import { selectLocale } from '@/redux/uiSlice'
 import { URI } from '@/types'
 import { Trans } from '@lingui/react/macro'
-import { Link } from 'react-router-dom'
+import { Link } from 'react-router'
+import encodeURIComponent from 'strict-uri-encode'
 import styles from './AccommodationInfo.module.scss'
 
 export const AccommodationInfo = ({